add version.txt to specify Node version
The motivation for this is that using the branch name is not
very practical when working on a staging branch where the
branch name does not match the version being developed for.

Instead the version would be specified/updated in version.txt
and then a pull request created against the staging branch
for that major version.