src: re-enable headers tar.gz
The generation of the headers tar.gz was commented out because after
applying the patch it failed as there were uncommitted changes. We now
commit the patched test and can proceed with generating the tar.
The commit is never used as we build in a detached head but that is
fine.