add support for 10.0.0-proposal
This commit updates the rpm spec file to node 10.0.0-proposal
and is the first commit that uses the OpenSSL version shipped with node.
This commit also adds a patch to update the version of node. This is
only required for a candidate release and will not be in the final
release.