build: support arm64
This commit is an initial commit for supporting ARM achitectures. The
goal of this commit is to discover any issues there might be and come up
with a configuration that builds, which can then be built out has needed.

Travis has also been updated but it remains to see if travis can handle
the longer build time. If not we should look into being able to split
the work up and only run one architecture at a time.

Refs: https://github.com/bucharest-gold/node-rpm/issues/138