src: add fips support
This commit add FIPS support for 8.x by producing an rpm which has been
compiled and linked to the fips module.

Currently the fips module is downloaded, but later we will probably add
this to our node fork and just have that unpacked. When that happens we
can remove the downloading from create_node_tarball.sh and that should
be the only change required.

There is an issue with this currently which is that the fips rpm cannot
have the same name for the node executable. Instread it is named
node_fips. It is up to the user of the rpm to create a logical link if
this name is not acceptable.