introduce quiet variable to spec file
This commit introduces a quiet variable to the spec file and if it is
set to true (1) then the complete output of the make test command will
be redirected to /dev/null. This is done as currently there will still
be output written to Travis log even if a silent build is requested.